+++ operator/runner/runner.go
@@ -579,12 +579,13 @@ func (r *Runner) poll(ctx context.Context) error {
 	if err == db.ErrOptimisticLock {
 		return nil
 	} else if err != nil {
-		logger.WithFields(
-			logrus.Fields{
-				"stage-id": p.ID,
-				"build-id": p.BuildID,
-				"repo-id":  p.RepoID,
-			}).Warnln("runner: cannot ack stage")
+		logger.WithError(err).
+			WithFields(
+				logrus.Fields{
+					"stage-id": p.ID,
+					"build-id": p.BuildID,
+					"repo-id":  p.RepoID,
+				}).Warnln("runner: cannot ack stage")
 		return err
 	}